Refactored to use internal packages
